Implements initial flow for creating and adding tracks to playlist.
* Updates addToPlaylist to take array

We were running into server errors when making the requests separate.
The Spotify API allows for sending an array of tracks which allowed for
us to avoid the server issues and add multiple tracks. This required the
use of promises to collect the track IDs via AJAX calls and pass them
all at once.

TODO:

- [ ] Split songs into groups of 100
- [ ] Find a way to maintain proper ordering
- [ ] Refactor playlist creation logic
- [ ] Fetch username based on authentication
- [ ] Dynamically create playlist name