Fixes issue displaying setlists with no state.
There are some cases where the venue doesn't have a state (e.g.
non-American countries). Since `Event#venue` uses `fetch` to get values
from the response hash and the response did not have the `@state` key
this would raise a `KeyError`.

This changes the logic to return the country name if the state is not
found.

We may want to consider not using fetch and retuning `nil`/empty strings
rather than having the user end up stuck on a loading page.