Extracts Event::Venue into its own class.
After handling the logic for when a Venue does not have a state it made
sense to change Venue from a struct into a class and handle parsing the
response in the class.

The use of `fetch` was also dropped. Since `fetch` would bubble up to an
exception I decided I would rather just have missing elements in the
interface and add satiable defaults if necessary.