Add support for searching setlists via Setlist.fm
Create `SetlistFm.API.ExternalClient` to communicate with Setlist.fm's
API. We are using this to search for setlists by artist name based on a
given query.

- We are using [Req](https://github.com/wojtekmach/req) as our HTTP client
- We are using [Bypass](https://github.com/PSPDFKit-labs/bypass) for
  API testing; allowing us to mock respones with pre-canned examples
- We should now be better leveraging Hammox for checking the typespecs
  of our `SetlistFm.API` behaviour (we had `import Mox` still in one
  place)

This also includes setting up the infrastructure for secret storage
because we now hav a secret (the Setlist.fm API key). We are following
the advice from [this post](https://felt.com/blog/elixir-configuration),
which is to use environment variables and set the Elixir application
config in `config/runtime.exs`. For local development, we are using
[dotenv_parser](https://hexdocs.pm/dotenv_parser/DotenvParser.html) to
read from a local (gitignored) `.env` file which will contain the
secrets.