Improve project set up process
Create scripts for setting up the project, including creating an `env`
file.

Taking loose inspiration from GitHub's
[scripts-to-rule-them-all](https://github.com/github/scripts-to-rule-them-all),
this commit includes scripts to ease the initial set up process and
running a local server. These scripts are mentioned in an updatd README.

As a part of making the setup process easier, this adds a
`.tool-versions` file, so we can rely on `asdf` for fetching the
languages we are using.