Adding caching to ExternalClient
Add cachex as a dependency and add basic caching for
`SetlistFm.API.ExternalClient` calls to the API.

This is a basic, first-pass implementation done to avoid hitting the API
while styling the search results page. It will likely require revisiting
to explore options such as cache size and handling expiration.