Create hook for checking if user authenticated
Copying the patterns set up in LiveBeats, define `UserAuth` module which
adds a hook that checks the session to see whether we have stored
anything about our Spotify users.

As a part of trying to make this nicer to work with, we defined a
`MusicAccount` struct that will include the username and access token,
allowing us to put the username in the nav header if the user has been
authenticated. This required making an API call to the `/me` endpoint of
the Spotify API.