Rework sign in flow
In order to track where to send a user back to after the log in, we add
a LiveView lifecycle hook to track the URI for non-authenticated users.
This is only set on the socket which we lose after going through the
OAuth flow, so, before redirecting a user directly to Spotify to auth,
we send them to a `/signin` route within the app, appending a
`redirect_to` query param that we can store on the session and use when
the user is brought back to our app after authenticating.

In addition to supporting this workflow, we also steal some code and
practices from the Phoenix auth generators and expand UserAuth to handle
signing in and out, including renewing sessions.