Remove `connected?` in `Setlists.ShowLive`
We had a `connected?` check in our `ShowLive` mount callback to avoid
hitting the API twice. Since we have implemented caching, we should no
longer hit the API twice (though we will hit the cache unnecessarily).
Since the cache check should be quick and it cleans up the code, we
moved the `connected?` check from `ShowLive` and fetch the setlist on
the first render.

As a part of this move, our expectation for a single call to our mock
API client was no longer true (it would be called twice now, once on
first mount and again on the socket connected mount). As an indication
that we didn't want to make two API calls our test `expects` explicitly
called out expecting it to be called once. Rather than updating this to
two for our test (which may be confusing) or removing the call count
altogether, we move the API caching out of the actual
implementation (`ExternalClient`) and into `SetlistFm.API`. This makes
our test/mock client behave more similarly to our actual client in that
it will also leverage a caching layer. This came at the cost of having
to have (Ham)mox work in "global mode"[^1] which means we can no longer
run these tests in async. Since the project is still small, there isn't
a noticeable cost to this.

[^1]: https://hexdocs.pm/mox/Mox.html#module-global-mode