fix: fix env setting for prod
`config/runtime.exs` has some code to use `DotenvParser` in dev and
test, but the `case` did not handle other environments. The expectation
is that production will not use a .env file and should have environment
variables set in other ways. This change updates the code to make other
envs be a no-op here.